Cut implicit modules: add modelVerifier test
Summary:
The changes from D37771613 (https://github.com/facebook/pyre-check/commit/5a14350a2381f6c4ea1248a5bc3f735577ef65d8)  introduced a bug in pysa that was not
caught by tests: when a model refers to a symbol in a deeply nested
module whose ancestors are not importable in post-D37771613 (https://github.com/facebook/pyre-check/commit/5a14350a2381f6c4ea1248a5bc3f735577ef65d8) pyre
semantics, the models don't resolve properly.

The underlying problem is that a block of code in pysa assumes all
parents of valid imports are themselves valid imports. This is true
in the python runtime and in old Pyre behavior, but in order to
support lazy module tracking we are modifying pyre semantics so that
only directories with python files as direct children are importable;
this prevents us from having to do a full recursive scan when
resolving imports (which is a no-go in large mixed-language repositories).

Reviewed By: grievejia

Differential Revision: D38073695

fbshipit-source-id: 66ef050e8c4530914a3c50496bc68f86c8e30792