cut implicit namespace modules: update legacy_resolve_exports behavior
Summary:
When we made the change in D37771613 (https://github.com/facebook/pyre-check/commit/5a14350a2381f6c4ea1248a5bc3f735577ef65d8) to reduce the number of implicit
namespace modules, it gave clean results in our QA runs.

But we've discovered that there are edge cases where re-exports aren't resolving
properly in fbcode projects, and furthermore pysa is showing model validation
errors in T126637425.

It appears that the underlying problem is, we didn't modify `legacy_resolve_exports`
to handle missing parent modules. This was easy to miss because (a) both the parent
and grandparent have to be missing for the bug to manifest, and (b) in distillery,
excessive re-exports are relatively rare.

This diff fixes the bug.

Reviewed By: grievejia

Differential Revision: D38054740

fbshipit-source-id: 99821acd9235adbff7eaace0966fe6afe9f1d46d