Cut implicit modules: add legacy_resove_exports test
Summary:
The change in D37771613 (https://github.com/facebook/pyre-check/commit/5a14350a2381f6c4ea1248a5bc3f735577ef65d8), which caused Pyre to stop allowing
import of directories with no python files as direct children,
led to a bug: When there are deeply nested modules (at least three layers of
nesting are necessary to produce the bug), the `legacy_resolve_exports`
function was failing to remap aliased names (such as re-exports). This caused
essentially no problems in the repo where we QA most changes like this, but in fbcode the nesting can be much deeper and it led to type errors
in several projects.

In both the bug was due to this part of the code assuming that all
parents of valid imports are themselves valid imports; this is true
in the python runtime and in the old module tracker, but it won't
work for a lazy module tracker because it requires recursively
scanning directories (which is something we cannot allow in huge
mixed-language monorepos).

Reviewed By: grievejia

Differential Revision: D38073696

fbshipit-source-id: c681a2f69dc3f5b12c7480e78f603b9f9b2b6bf5