ModuleTracker: s/module_to_files/qualifier_to_modules
Summary:
I don't really want to start rewriting the internals of module tracker using
the existing variable names, some of them I find confusing.

`module_to_files` is less clear than `qualifier_to_modules` because the latter
is much better aligned with mapping an import name (i.e. a qualifier, which
we currently represent as a Reference.t) to a list of ModulePaths.

Reviewed By: grievejia

Differential Revision: D37956290

fbshipit-source-id: f7e72fef3adc3672d66e5830dc7d1614ecde7c2b