ModuleTracker: move find_module_paths up
Summary:
I want to factor parts of `find_module_paths` into functions that can eventually
be reused by both lazy and eager module tracking. These modules are going to need
to live above the `Base.Layouts` module definition, so moving the function up
is necessary.

It's also good for readability of the code in general, since `find_module_paths`
runs before the `Layouts` code at runtime anyway - it's now easier to skim the
code top-to-bottom.

NOTE: everything in this stack from here on is working toward D37972795, which changes the (still nonlazy) module tracker to handle implicit modules using a data structure that we can adapt to lazy tracking.

Reviewed By: grievejia

Differential Revision: D37958435

fbshipit-source-id: 4dfc2b5a333a0c046d2e0d2c8f0020f2d691bd72