Add function_name field to client
Summary:
### Context:
We want to add a dashboard for top unannotated library functions, on both the project and overall level in `fbcode`.

Design doc is [here](https://docs.google.com/document/d/1JIztqkOSEONNPluRKZKiiUeV9VsE_VtSqT15qch7Xmc/edit#).

### Problem:
We want to client to recognize the `function_name` field that the backend sent.

### Solution:
We update the `CoverageGap` class

### Stack:
- Send whether something is library function and an optional function name field
- (This) Update client
- (Next) For each path, log function names and counts into scuba table
- Extract data from table and display on frontend

Reviewed By: pradeep90

Differential Revision: D38054408

fbshipit-source-id: 8943a327c2b78ef5f4f4f68ff29fd250a6190b95