Send optional function_name with CallableReturnIsAny
Summary:
### Context:
We want to add a dashboard for top unannotated library functions, on both the project and overall level in `fbcode`.

Design doc is [here](https://docs.google.com/document/d/1JIztqkOSEONNPluRKZKiiUeV9VsE_VtSqT15qch7Xmc/edit#).

### Problem:
We need a way to identify whether something is a library function, so that we can log it and display this on the dashboard. We also need to know the name of the library function that we are logging.

In this diff we'll address functions with unannotated return types, in the next we'll look at functions with unannotated parameters.

### Solution:
As these library functions are function calls, when an expression has type any and is a call, we know that this is a library function. So, we split the `TypeIsAny` case into another `CallIsAny` case. We also obtain the function name from the expression itself and send an optional `function_name` field in the coverage gap.

### Stack:
- (This) Send whether something is library function and an optional function name field for functions with unannotated return types
- (Next) Update client
- Send function name for library functions with unannotated parameters
- For each path, log function names and counts into scuba table
- Extract data from table and display on frontend

Reviewed By: pradeep90

Differential Revision: D38115043

fbshipit-source-id: be4434830294018db7d3d8f305bcd067fe7432fa