Send function name for CallableParameterIsUnknownOrAny coverage gaps
Summary:
### Context:
We want to add a dashboard for top unannotated library functions, on both the project and overall level in `fbcode`.

Design doc is [here](https://docs.google.com/document/d/1JIztqkOSEONNPluRKZKiiUeV9VsE_VtSqT15qch7Xmc/edit#).

### Problem:
In the last diff we got function names for functions with unannotated return types.

Now, we want to send the function name for expressions with unannotated parameters, which is the `CallableParameterIsUnknownOrAny` scenario.

### Solution:
For `CallableParameterIsUnknownOrAny`, we extract the attribute name and send this function name to the client.

### Stack:
- Send whether something is library function and an optional function name field for functions with unannotated return types
- Update client
- (This) Send function name for library functions with unannotated parameters
- (Next) For each path, log function names and counts into scuba table
- Extract data from table and display on frontend

Reviewed By: pradeep90

Differential Revision: D38058240

fbshipit-source-id: 3e5a84412ceef3c20148cb364cb53bceeee49ecf