Add CallableReturnIsAny coverage gap for unannotated return type
Summary:
### Context:
We want to add a dashboard for top unannotated library functions, on both the project and overall level in `fbcode`. We want to obtain function names for functions with unannotated return types and unannotated parameters.

### Problem:
Functions with unannotated return types include callable expressions with the `Any` type, which we currently capture as a coverage gap, but this information does not allow us to obtain the module name.

Hence, we need to add a coverage gap for callables with the `Any` parameter, and then later extract the function name as the type contains this.

### Solution:
We add the `CallableReturnIsAny` type. Note that if a coverage gap has both unannotated returns and parameters, the check labels it first as the `CallableReturnIsAny` type instead of the `CallableParameterIsUnknownOrAny` when it can potentially be both.

### Stack:
- Add coverage gap for unannotated return type
- Add function name for unannotated return type
- Add function_name field to client
- Send function name for Callable coverage gaps
- Log to scuba, hookup to hive
- Display on frontend

Reviewed By: pradeep90

Differential Revision: D38081949

fbshipit-source-id: e0cc74e0d2fc567cf67ff890c0d5cb7249940384