ModuleTracker: Use the same filtering for incremental update and creation
Summary:
The existing ModuleTracker code is inconsistent in how it handles
module filtering for creation vs updates
- we explicitly filter creation, which includes dotfile filtering
- we only implicitly filter updates relying on ModulePath.create
  logic that is (a) inconsitent and (b) something I'm hoping to
  remove

This caused an incremental update bug that I illustrated in the previous commit.

Here, I refactor ModuleTracker to use the same filtering logic everywhere which
kills the bug. In the next commit I'll remove the filtering logic from ModulePath,
which shouldn't be relied on (single responsibility principle says ModuleTracker
ought to handle all path-name-based-filtering if it handles any)

Reviewed By: grievejia

Differential Revision: D37963879

fbshipit-source-id: 7665fdffee04fec264bac9f6402937c0bafeca05