Make a single create function
Summary:
In order to make lazy module tracking work, I want to have a new type
that represents just the location information (relative + priority) of a
ModulePath, but not the information (like is_stub, etc) that is specific
to regular python files.

The idea is that `ModulePath.create` should rely on this new type's create
function; the current logic where `create_from_search_path` (which is a
misnomer anyway, it's really "create from search path*s*") doesn't play
well with that refactor.

Let's just put it in `create`; the logic isn't complex enough to really need
its own function with a distinct interface, and it will make later refactors
simpler.

Reviewed By: grievejia

Differential Revision: D37963877

fbshipit-source-id: c3e48dcc0680f46d02113fd71f853601d4577ccb