Simplify ModulePath: remove redundant extension checks
Summary:
In the previous two commits I exposed and fixed a ModuleTracker bug
caused by us trying to rely on ModulePath to do extension filtering; this
isn't a good idea, the filtering should always be done in one place and
ModuleTracker (which handles navigating the filesystem) is the logical place.

So let's get rid of the extension filtering from ModulePath, the code is
more confusing and less efficient when we have redundant checks.

Reviewed By: grievejia

Differential Revision: D37963875

fbshipit-source-id: c13bb54cc44ffd403e0b59dac862613957fe42b9