ModuleTracker: factor out filesystem_events logic
Summary:
In order to rearrange the data in a way that will work for lazy module
tracking, the implicit modules need to exist in a multiverse-aware structure
just like explicit modules (by this I mean potentially multiple entries
due to the same relative path appearing in multiple search roots).

This means we'll need to consume raw filesystem updates; this requires
that we factor the logic into a stand-alone helper function first.

Reviewed By: grievejia

Differential Revision: D37972794

fbshipit-source-id: 0eecd55236d0e306b4cee7faed0e27d4eaac05d5