Add ModulePath.Raw to represent an artifact path, mapped relative to a root
Summary:
In order to make lazy module tracking work, I need to be able to work with a
raw representation of ArtifactPath that is similar to ModulePath.t but is suitable
for "raw" paths that might not refer to python files; I'll use this in my representation
of directories.

This also allows us to modularize the raw-path portion of ModulePath creation (which
involves the code for mapping a path as relative to a given search path) from the
other parts that involve working with extensions, etc to create explicit-module-specific
fields.

Reviewed By: grievejia

Differential Revision: D37963878

fbshipit-source-id: e8164edcc19b71aea7da6fb1211dfaf1d4026da5