ModuleTracker: add a test illustrating a bug
Summary:
While trying to simplify ModulePath and deduplicate logic in
order to make lazy module tracking easier, I realized we have a
bug in incremental module tracking:
- we explicitly use filter logic at the ModuleTracker layer in
  creation
- but we don't filter in updates, instead we implicitly rely on
  ModulePath creation returning None, but the codebases are not
  the same

This illustrates why we shouldn't be doing *any* pathname-based
filtering in ModulePath - we should have only one way to filter
and only one ocaml module that does the filtering, it reduces the chance of inconsistency in addition to minimizing code duplication.

I'll fix the bug in the next diff by having ModuleTrcker handle
the filtering, and after that remove the undesirable ModulePath
logic.

Reviewed By: grievejia

Differential Revision: D37963880

fbshipit-source-id: 3c4b34eee447e591843e89f1903ff033c031c526