ModulePath: move create_from_search_path
Summary:
In order to unblock lazy module tracking, I need to split parts of the
ModulePath logic into a ModulePath.Raw module that is suitable for logic that
is relevant to both directories and files versus the main logic that is only
relevant to python files.

That change itself is messy given the way we've split create and
create_from_search_path, so I want to first clean up the existing code.

This commit preps for cleanup by moving the functions I want to modify
to live next to one another (which will make the next diff easier to review)

Reviewed By: grievejia

Differential Revision: D37963876

fbshipit-source-id: a7dcba8f237f27e59e0afab3201a56fda56be3e8