ModuleTracker: make FileSystemEvent contain a ModulePath
Summary:
In order to move module tracker toward a data model that we can use for lazy
tracking, I need to track each specific ArtifactPath of a folder that corresponds
to an implicit namespace package module.

This will require that the updates come in already processed into ModulePath form,
which is easy to accomplish: just convert the path in FileSystemEvent.create so
that we produce an option and use `filter_map` in `create_filesystem_events`.

An additional upside of this refactor, besides that it unblocks moving toward
lazy tracking, is that the amount of nesting in `process_filesystem_event`
goes down significantly.

Reviewed By: grievejia

Differential Revision: D37972797

fbshipit-source-id: 8383d1eaabc4d7964a6d58af560334630e744a5c