Turn result tuple into a record.
Summary:
Problem: Hard to understand why it was doing what it was doing.

The original diff had a lot of discussion.

(I will probably change this going forward. I don't think we need this.)

This diff: Just change the tuple to a record (for easier review).
Next diff: Change the field name.

Reviewed By: grievejia

Differential Revision: D38136253

fbshipit-source-id: b096d818fc502855b146cc0ea9bea68b8f5c071a