Add `PyrePath.read_directory_ordered`
Summary:
This logic already existed, as we use it in the crawl to build
ModuleTracker. But we need it to be exposed in the public API
directly in order to implement lazy tracking; the current
approach of only exposing it indirectly via listing ordinary
files won't work for lazy tracking.

Reviewed By: grievejia

Differential Revision: D38170465

fbshipit-source-id: cdf58a5b3a91ea5761b2cf64bca0c5be29f64bab