Add test of handling non-utf-8 data in async TextReader
Summary:
Since we know that other encodings can sneak in (we've had previous bugs
where, for example, string literals in alternately-endoded files cause ocaml
to send us non-utf-8 data) this could cause a SEV if a type error in a non-utf-8
file is committed; let's fix it.

This adds a test verifying that our async server connection will crash
if ocaml sends non-utf-8 data to us; the next commit will patch the bug.

NOTE: I'm not actually sure how to do an async assertRaises using unittest /
testslide; in pytest there's a context manager we can use for this but the best
solution I could find for our tests was a manual try/except block.

Reviewed By: grievejia

Differential Revision: D38134054

fbshipit-source-id: 461d895285a9092b887338e33a278358935337b4