s/FileSystemEvent.Update/FileSystemEvent.NewOrChanged/g
Summary:
In the spirit of namespacing logic and consistency, I want to change
the name of `FileSystemEvent` to `FileSystem.Update`, which will
improve the consistent organization of ModuleTracker code.

But `FileSystem.Update.Update` is a very awkward name for a variant;
and in reality this just illustrates that the name isn't as clear
as it should be anyway. The fact that this layer of the code actually
cannot distinguish between new and changed is very important, it
impacts what kinds of designs are feasible for us. So using the
more descriptive name `NewOrChanged` makes the code clearer.

Reviewed By: grievejia

Differential Revision: D38225333

fbshipit-source-id: a6dbe02ad9b7d28b22a3589ed645a084080b564b