Move ModulePaths, ExplicitModules, ImplicitModules up
Summary:
When I created these namespaces they just had data types in them, and it
made sense to me to put them next to `Layouts.t` where it was easy for me
(and code reviewers) to see how the types fit together.

But now that I'm moving logic outward, it's clearer that these types aren't
actually coupled to anything in `Layouts`, or even in `Base`, and it makes more
sense for them to just be at the top level.

Reviewed By: grievejia

Differential Revision: D38226343

fbshipit-source-id: 86ca7caba76d9936081d2ebfaf0c377a22821830