ModuleTracker: flip over to using a table of implicit locations
Summary:
This commit flips us away from using submodule_refcounts - which is tighly
coupled to eager module tracking - toward using a laziness-friendly layout
where each implicit path (there can be multiple in different search
roots!) is tracked independently.

This is a key unblock for lazy tracking because we need a format that
allows us to determine whether to allow import of a namespace package
without crawling the full tree underneath it.

Reviewed By: grievejia

Differential Revision: D37972795

fbshipit-source-id: 4c53bad828f50d3923bcde11d1cf96148b4e4277