Move the merge_updates logic
Summary:
Now that we have namespaces representing raw files,
expicit modules, and implicit modules it is time to start
moving logic that has a clear owner to the right place.

The `merge_updates` function is quite a bit of code, but it is actually very simple and has no interaction with
the changes needed for lazy-vs-nonlazy module tracking.

Since it is a pure function that just deduplicates
`ExplicitModules.Update` events, let's move it to that
namespace.

Reviewed By: grievejia

Differential Revision: D38225327

fbshipit-source-id: 831dcc9165930718f71c04ceaf554179d0dcba24