Reorganize types into namespaces
Summary:
In order to make a final decision for how to move forward with
lazy module tracking, I need to first get a better idea what
a well-modularized API looks like.

The reason is that it probably makes sense to separate lazy
and nonlazy trackers *if* I can share most of the logic. But
duplicating the logic would put us in a bad state, and it would
be better to use mutable state for nonlazy updates if duplicating
many difficult-to-write and difficult-to-test lines of logic
is the only option for separating the implementation.

As a first step toward modularizing the ModuleTracker logic,
I'm creating namespaces into which logic can go. As of this
commit, the namespacing is only applied to data types, but I
will be moving functions into the appropriate namespace in
later commits.

Reviewed By: grievejia

Differential Revision: D38225329

fbshipit-source-id: e3b163b82e557514ca054d52c7832ee778cc262d