ModuleTracker: rearrange some implcit logic
Summary:
This commit kills the ModuleFinder.Implcits module, moving
all required functionality into ImplicitModules so that it can
live closer to where it is used.

In the process, I realized I could get rid of the indirection
represented by the `Existence.change` type, which was only needed
because the previous ordering of the code meant we didn't have access
to the appropriate update type at the type we were defining `to_change`.

Reviewed By: grievejia

Differential Revision: D38225331

fbshipit-source-id: 1c1ead797b498132f2bc5589bd6bf6d901120fd8