Rename the layouts tables in ModuleTracker
Summary:
In keeping with the new namespaces ExplicitModules and
ImplicitModules, let's use explicit_modules and
implicit_modules as the field names for the
`ExplicitModules.Table.t` and `ImplicitModules.Table.t`
values inside a `Layouts.t`.

Better names will make it easier to hold the code in our
head and refactor until laziness becomes easy to achieve.

Reviewed By: grievejia

Differential Revision: D38225334

fbshipit-source-id: ed44981dcfb118cf560302b8cb0f2b019d7a5c8e