ModuleTracker: push some logic to Layouts.Api
Summary:
A previous commit in this stack pushed some very ModuleTracker-specific
logic (e.g. lookup_module_path, is_implicit_module, all_modules) down
to the level of the `Table.Api` layers because it was important at that
stage to isolate these read APIs.

But now that we have a `Layouts.Api` abstraction (which was missing at that
stage) it makes sense to define these functions in terms of table-level
operations and move them to that module.

They can all be defined in terms of `find` and `data`; the former should
be relatively easy to write in a lazy way but the latter cannot be lazy.

Reviewed By: grievejia

Differential Revision: D38235012

fbshipit-source-id: ee3544cdb42f91849da79a9e021b36f314018139