Add read operations to Table.Api interfaces
Summary:
In order to properly abstract out the Table api, we need to include
all the read-only operations on tables as part of the record of callbacks
so that no downstream code relies directly on the implementation.

At this stage we're still reaching across layers in `Base.read_only`
because the Layouts module hasn't yet been split into an Api versus
a concrete implementation; that will be in the next commit.

Reviewed By: grievejia

Differential Revision: D38235017

fbshipit-source-id: 981fa0c35c31abe1367289edd7928aa988b6c011