Move in-memory source handling to EnvironmentControls
Summary:
It has for some time been a thorn in our side, when dealing with the
environment stack, to have to maintain both `create` and `create_for_testing`
functions at every layer.

At this point it has become a major issue because it is forcing us to do
work at the ModuleTracker.Base level that we need to push down into Layouts.
Let's just fix the problem by making in-memory sources an option in EnvironmentControls.

This groups together the various flags that control environment behavior (as well
as assertions needed to deal with the fact that some settings disable certain
operations) in one place.

Reviewed By: grievejia

Differential Revision: D38265764

fbshipit-source-id: 33e2512dcfbddb737c42fe4ec04730976034137c