Split Table apis
Summary:
In this commit, I begin working toward abstracting the
table api needed for updates (which are by far the trickiest
logic in ModuleTracker, and therefore the part that I desperately
want to avoid duplicating for lazy vs nonlazy trackers) from the
concrete represnetation of tables as hash tables.

I do so by creating `Table.Api` modules that represent tables as
records of callbacks and having the update functions (now renamed
to `update_module_paths` which is both clearer and avoids a name
collision with the `Table.update` function from the table api)
work against that api.

The concrete representation has been moved into `Table.Eager`
and that is what `Layouts.t` relies on.

Reviewed By: grievejia

Differential Revision: D38235014

fbshipit-source-id: c93e328d1c6223f46efbe996b2f69d5944ac8657