Rearrange ModulePaths module
Summary:
It is time to begin lazy module tracking work. This will require
adding new submodules across ModuleTracker, including in the ModulePaths
layer.

As a result, it's important to make sure code is in a good order
and well namespaced; the `find_all` function relates to basic module
path finding and belongs in the `Finder` module for now (debateably
we should make an `Eager` module instead, but that seems like overkill)
whereas the logic for creating updates from artifact paths belongs in
the `Update` module.

Reviewed By: grievejia

Differential Revision: D38276764

fbshipit-source-id: d740ba2e2c12ae5a50d9cb104e44f27874607bf0