ModuleTrackerTest: reorder
Summary:
The `test_search_path_subdirectory` is actually only verifying ModulePath
logic, there's no use of ModuleTracker. Move it to the top.

I considered splitting these tests into a ModulePathTest module, but the truth
is that we do save a lot of code duplication by keeping all these tests in one
file so I'll leave it for now.

Reviewed By: grievejia

Differential Revision: D38306931

fbshipit-source-id: b646a5edbb03d21b4b2bf4c7d1b6cebb104c8924