Bugfix: fix PyrePath.read_directory_ordered
Summary:
I added read_directory_ordered mostly by accident; it was needed
in an earlier lazy module tracker commit that I later backed
away from to focus on refactors.

As a result, I forgot to unit test it, and didn't notice that I
was treating the output of `read_directory_ordered_raw` as
if the output was as absolute paths, when in reality it is as
relative paths.

Reviewed By: grievejia

Differential Revision: D38293750

fbshipit-source-id: 9174053a48a3f9da87349d3e85da5a790dad1566