ModuleTrackerTest: break helper functions out
Summary:
The module tracker test code is weird, we have about 12 tests
all grouped together into a massive, 2000 line test called
`create_test`.

This is not great - one problem is that it makes skimming the code
hard (for example vim paragraph navigation is usesless because of
a lack of line breaks) but it also means we can't easily isolate
tests using the `-only-test` flag in `dune exec`.

So I'm going to split it up. The first step is to get the helper
functions extracted; I also decided to rename `assert_create_fail`
to `assert_no_module_path` which I think is more descriptive.

Reviewed By: grievejia

Differential Revision: D38306927

fbshipit-source-id: b5c76925036a6a061e20a20bf0cbc604a743caae