Fix same_module_compare semantics
Summary:
The word `compare` has a preexisting meaning in ocaml
sorting algorithms: things sort from smallest to biggest.

We shouldn't be reversing the semantics but we are; I remember
being very confused by the code because higher-priority
`left` values wind up returning positive rather than negative
values which is backward.

We didn't notice this because we use hand-coded sorting logic,
but for lazy module tracking there's going to be a step where
we ought to call List.sort and it is very confusing if we
get the ordering backward.

(Note that we no longer have to invert Int.compare, or use
right - left... these were all code smells that our semantics
were backward!)

Reviewed By: grievejia

Differential Revision: D38306928

fbshipit-source-id: 67e25a1819c11d4099bf7566da37a106f5388f1b