ModuleTrackerTest: move and rename all non-tracker tests
Summary:
I became even more aware, as I worked through testing the lazy module
tracker, of the fact that we have pure ModulePath logic tests intermixed
with ModuleTracker tests.

This situation may not be ideal but I can see arguments in favor of it;
ModuleTracker behavior depends heavily on ModulePath so keeping the
tests together is handy and it allows a lot of fixture code sharing.

But I do think we should at least make it very clear by having all the
ModulePath tests grouped together at the beginning and naming them
consistently. That's what this commit does (previous commits had
attempted to make some progress on this but I hadn't realized how
disorganized the code was yet).

Reviewed By: grievejia

Differential Revision: D38306929

fbshipit-source-id: 00a316f50c924078313f9b28f0f313d8aa9efbb2