Lazy ModuleTracker logic: LazyFinder
Summary:
This commit (split out of a larger commit to make review
easier) implements the core directory logic that will
power a lazy tracker.

The functions are temporarily exposed in the .mli file to
prevent ocaml from complaining about unused code, but they
will be private in the final implementation.

They have no unit tests because - consistent with the rest
of this module - we test all logic only in ways that rely
in the public interface of ModuleTracker; the existing
read-only logic tests will provide quite a lot of coverage
that this code works properly once everything is wired up.

Reviewed By: grievejia

Differential Revision: D38366388

fbshipit-source-id: a6dff8ffe490c0d1d5ce5a4b95155e6831325e56