Lazy ModuleTracker test harness [1/2] read-only tests
Summary:
Based on diff review requests, I'm trying to split out the test
logic for lazy module tracker in a way that optimizes for easy
review. This commit tweaks all of the non-update tests that do
not rely on exhaustively listing modules so that it will be easy
to run them against both lazy and nonlazy module trackers.

For now, the `run_lazy_and_nonlazy` helper only actually runs
tests against a nonlazy tracker (since the lazy tracker does not
yet exist) but it will be a one-line change to turn the tests on.

The remaining tests all require listing modules, which only an eager
tracker can do.

Reviewed By: grievejia

Differential Revision: D38366391

fbshipit-source-id: c111385553f89e7fdbab66c024c2b2e70e478592