Lazy ModuleTracker test harness [2/2]: new update test
Summary:
Because the existing module tracker update tests depend very
much on eager tracking, it does not make sense to adapt them
for lazy tracking the way I did with most of the read-only tests.

Instead, we'll need a new test that covers the laziness-specific
edge cases; fortunately for us, the vast majority of update logic
is the same for lazy and eager environments so we don't need
to duplicate most of the update test cases (there are a lot of them!).

We don't yet have a lazy tracker, so what this commit does is
prepare the test harness by writing out each of the test cases
we *will* use against a nonlazy tracker.

As a result this commit puts us in a weird state where we have an
unnecessary test whose name is misleading, but what I'm optimizing
for here is the ability to review the critical logic change
by seeing *exactly* what actually changes.

Reviewed By: grievejia

Differential Revision: D38366389

fbshipit-source-id: bbd9afe75b96da778f55d88358970c7c6efeb974