Lazy ModuleTracker logic: LazyTracking table abstraction
Summary:
Much like the previous commit, this commit breaks out a layer of
logic we need for lazy module tracking into a separate commit.

This code will all be private in the final implementation, but we
have to temporarily expose it in the .mli file to prevent the
compiler from complaining about dead code.

The behavior will be tested indirectly using the public module
tracker API, so there are no unit tests of this layer; in the next
diff we'll actually implement the lazy tracker and test everything.

Reviewed By: grievejia

Differential Revision: D38366390

fbshipit-source-id: 41f443d4b707ce0e4109ee15daa2c0fcf7176ce6