Bugfix for localized references [1/2]: add tests
Summary:
It is possible for us to look up the names of local variables
directly during type checking; this is not something we initially
realized, and so `Reference.possible_qualifiers` is incorrectly
implemented in a way that fails for locals.

This commit adds tests demonstrating the bug; the next will fix
it.

## Why does it matter?

It turns out that this causes a long-standing but invisible bug in lazy
parsing: it is possible for defines and classes to be local (and
therefore name-mangled) while still being part of UGE.

When this happens, Reference.possible_qualifiers will produce an
empty list, and therefeore the lazy-loading in UGE will fail.

It is basically impossible to actually trigger this bug because
you'd have to directly look up a local define, and we only know
about them by calling `all_defines_in_module` (which will correctly
lazy-load everything).

But the same bug also affects the unsaved changes ownership checks
in ModuleTracker; Josh discovered this as he was implementing a full
integration test suite based on the existing one for incremental check.

In this case, what happens is that we correctly get the local nested
function when we call `all_defines_in_module`, but then the
`typeEnvironment` ownership check causes us to skip type checking, so
that we fall back to whatever type errors were in the underlying parent
environment.

Reviewed By: grievejia, pradeep90

Differential Revision: D38397277

fbshipit-source-id: d14c400d9476b79b44774095193f5c46316ce6e8