OverlaidEnvironment: Do not log the number of errors
Summary:
We currently force all errors on the entire project inside of
overlaid environment updates, purely for the purpose of logging
the count.

This may actually be contributing to slow incremental updates, I'll
try to get some perf numbers on the impact of this diff.

But more importantly, it's blocking me from getting new perf numbers
because:
- in a lazy server, this stage causes incremental updates to freeze;
  at the very end of the update we trigger a *single-threaded* attempt
  to compute all the errors in the project! This is not ideal.
- in a lazy module tracker it just crashes the server, since listing
  all project qualifiers is illegal

We aren't even logging this data to scuba, so nothing is depending on
it; I didn't get rid of it when I first wrote OverlaidEnvironment because
I did not want to change any observable behavior but now we really
need it to go away.

Reviewed By: grievejia

Differential Revision: D38431685

fbshipit-source-id: 0b184fc267a147dac3e90746828127aaf2fcb6be