Add a helper method to find_directories_test
Summary:
In the next commit, I'm going to add more abilities to the typeshed
root finder so that it is possible to use a compacted typeshed
where the third-party stubs have been flattened into a single search
root.

At that point, the tests become pretty repetitive, so first let's
refactor so that the test cases are much more compact.

While we're at it, let's also have the stubs directory use a layout
more like an actual stubs directory, where the packages are nested
inside an extra outer layer. This doesn't make much difference
to the test logic, but it will make the intent clearer when we add
logic and tests for a more efficient layout.

The actual structure of typeshed is as follows:
- the stdlib directory is a search root, with stdlib stubs
- each directory inside of `stubs` is a search root, inside
  of which there are stubs for a single library

But our current logic is furthermore making every subdirectory
of every libary a search root! This isn't correct; for example
`stubs/urllib` is a search root, but `stubs/urllib3/urllib3`

The problem appears to be that `Path.iterdir` does not
do what the docs claim. See the extended test.

Reviewed By: grievejia

Differential Revision: D38503528

fbshipit-source-id: de7c0d4a889d4eac562beb7ed18a258c31c3a247