Clean up ModuleTracker: rename process_update to skip_update
Summary:
The name `process_update` in the Table.Api record was intended as a
predicate, as in `should_process_update`, but in code review we caught
that it actually sounds like a processing function (in fact, there are
two such functions already in this module!)

Let's rename it to avoid confusion. After some thought I decided that
a predicate named `should_skip_update` probably makes the code clearer
than one named `should_process_update` (especially since the if condition
was always `not process_update` anyway).

Reviewed By: grievejia

Differential Revision: D38389700

fbshipit-source-id: 51c182ae2018d92ed2d07deed9e164e71b5189cd