Rearrange download_typeshed.py
Summary:
I was finding it a bit hard to navigate the code; the data
was at the top in an inconsistent order and the functions were
also mixed together so that it wasn't easy to see at a glance
each operation in the pipeline (downlaod, trim, patch).

Since I need to add another step (combine stubs, so that we can
get more efficient reads at runtime) I very much want to clean
this up. To this end, I move the logic for each stage that has
a class into methods of that class, which results in namespacing
and indentation structure that reveal the logical layout of
the code.

I also renamed the types to make them shorter and more aligned
with the variable names.

Reviewed By: pradeep90

Differential Revision: D38507081

fbshipit-source-id: 962704a6a281e8ce170e985cd72fdc8bd4b0d033