Do not permit failed patches in typeshed downloads
Summary:
Even with slightly noisier logging, the typeshed download script
is currently too dangerous to run - it silently skips failed patches,
which is not an acceptable behavior.

Doing this just leads to CI failures (if the failed patches are going
to be exposed via our QA jobs) or worse yet a broken release (if we ever
need a patch that *isn't* exposed in QA it would very easily slip
through).

We have the patches because we expect them to be applied. If they fail
to apply, whoever is upgrading typeshed should either remove them
(if they are no longer needed) or fix them, and we should force this.

If for whatever reason we want to debug against a typeshed where failed
patches are skipped it is easy enough to tweak the script as needed.

Reviewed By: joshuaxu1995

Differential Revision: D38506576

fbshipit-source-id: 3b8757e73d8a133bc652fc820e1d785b920d2e7f