Remove useless patch
Summary:
When I was running the `download_typshed.py` script, I noticed this in the output:
```
[2022-08-08 09:10:50,963][INFO]: Applying patch scripts/typeshed-patches/stdlib/os/__init__.patch
patch: **** Only garbage was found in the patch input.
```

The issues is that that patch file isn't in a format we can actually apply.

This further means that none of our typeshed releases, at least in recent memory,
have actually applied this patch. That presumably means it isn't really needed!
Let's get rid of it.

Reviewed By: joshuaxu1995

Differential Revision: D38506575

fbshipit-source-id: edbc96215e7521753815659bd97f659bade91c0c