Proper join of ModeSet w.r.t. `Obscure` for user models
Summary:
This diff addresses the problem demonstrated in the previous diff. In
particular, for user models, we meet (instead of join) mode `Obscure` in two
`ModeSet`s (which is a set of modes). That is, if a user model contains mode
`Obscure` and another user model does not, then we want their joined model to not
contain mode `Obscure`.

An alternative is to redefine the join operation of two ModeSets (that is used
when joining two models -- not just two user models), but such definition also
requires updating the less-than-or-equal-to comparison operation between two
ModeSets, such that the result of joining two ModeSets is "larger" in the
lattice.

Reviewed By: arthaud

Differential Revision: D38541776

fbshipit-source-id: 6032a33218bccc2d106ebada622104645c555477